New module for adaptation functions
