+++ src/jamaica_infrastructure/analysis/utils.py
@@ -2,7 +2,6 @@ from typing import Any
 
 import pandas as pd
 from scipy import integrate
-import math
 import numpy as np
 from tqdm import tqdm
 
@@ -77,136 +76,3 @@ def risks(
     )
 
     return dataframe[index_columns + [expected_risk_column]]
-
-
-def prod(val):
-    res = 1
-    for ele in val:
-        res *= ele
-    return res
-
-
-def calculate_growth_rate_factor(
-    growth_rate=5.0,
-    start_year=2020,
-    end_year=2050,
-):
-    """Set discount rates for yearly and period maintenance costs
-
-    Parameters
-    ----------
-    discount_rate
-        yearly discount rate
-    growth_rate
-        yearly growth rate
-
-    Returns
-    -------
-    discount_rate_norm
-        discount rates to be used for the costs
-    discount_rate_growth
-        discount rates to be used for the losses
-    min_main_dr
-        discount rates for 4-year periodic maintenance
-    max_main_dr
-        discount rates for 8-year periodic maintenance
-
-    """
-    growth_rates = []
-
-    if isinstance(growth_rate, float):
-        for year in range(start_year, end_year + 1):
-            growth_rates.append(
-                1.0 * math.pow(1.0 + 1.0 * growth_rate / 100.0, year - start_year)
-            )
-    else:
-        for i, (year, rate) in enumerate(growth_rate):
-            if year > start_year:
-                growth_rates.append(prod([1 + v[1] / 100.0 for v in growth_rate[:i]]))
-            else:
-                growth_rates.append(1)
-
-    return np.array(growth_rates)
-
-
-def calculate_discounting_rate_factor(
-    discount_rate=4.5,
-    start_year=2020,
-    end_year=2050,
-    maintain_period=4,
-    skip_year_one=False,
-):
-    """Set discount rates for yearly and period maintenance costs
-
-    Parameters
-    ----------
-    discount_rate
-        yearly discount rate
-    growth_rate
-        yearly growth rate
-
-    Returns
-    -------
-    discount_rate_norm
-        discount rates to be used for the costs
-    discount_rate_growth
-        discount rates to be used for the losses
-    min_main_dr
-        discount rates for 4-year periodic maintenance
-    max_main_dr
-        discount rates for 8-year periodic maintenance
-
-    """
-    discount_rates = []
-    maintain_years = np.arange(start_year + 1, end_year + 1, maintain_period)
-    for year in range(start_year, end_year + 1):
-        if year in maintain_years:
-            discount_rates.append(
-                1.0 / math.pow(1.0 + 1.0 * discount_rate / 100.0, year - start_year)
-            )
-        else:
-            if skip_year_one is True:
-                discount_rates.append(0)
-            else:
-                discount_rates.append(1)
-
-    return np.array(discount_rates)
-
-
-def extract_growth_rate_info(
-    growth_rates, time_column, rate_column, start_year=2000, end_year=2100
-):
-    growth_year_rates = []
-    growth_rates_times = list(sorted(growth_rates[time_column].values.tolist()))
-    # And create parameter values
-    for y in range(start_year, end_year + 1):
-        if y in growth_rates_times:
-            growth_year_rates.append(
-                (
-                    y,
-                    growth_rates.loc[
-                        growth_rates[time_column] == y, rate_column
-                    ].values[0],
-                )
-            )
-        elif y < growth_rates_times[0]:
-            growth_year_rates.append(
-                (
-                    y,
-                    growth_rates.loc[
-                        growth_rates[time_column] == growth_rates_times[0], rate_column
-                    ].values[0],
-                )
-            )
-        elif y > growth_rates_times[-1]:
-            growth_year_rates.append(
-                (
-                    y,
-                    growth_rates.loc[
-                        growth_rates[time_column] == growth_rates_times[-1], rate_column
-                    ].values[0],
-                )
-            )
-
-    return growth_year_rates
-
